+++ src/Context/WordPressAdminContext.php
@@ -18,7 +18,6 @@ use \StephenHarris\WordPressBehatExtension\Context\Page\AdminPage;
  */
 class WordPressAdminContext extends RawMinkContext implements Context, SnippetAcceptingContext
 {
-    use \StephenHarris\WordPressBehatExtension\StripHtml;
     use \StephenHarris\WordPressBehatExtension\Context\PostTypes\WordPressPostRawContext;
 
     public function __construct(AdminPage $adminPage)
@@ -67,54 +66,8 @@ class WordPressAdminContext extends RawMinkContext implements Context, SnippetAc
      */
     public function iGoToMenuItem($item)
     {
-
-        $item = array_map('trim', preg_split('/(?<!\\\\)>/', $item));
-        $click_node = false;
-
-        $menu = $this->getSession()->getPage()->find('css', '#adminmenu');
-
-        if (! $menu) {
-            throw new \Exception("Admin menu could not be found");
-        }
-
-        $first_level_items = $menu->findAll('css', 'li.menu-top');
-
-        foreach ($first_level_items as $first_level_item) {
-            //We use getHtml and strip the tags, as `.wp-menu-name` might not be visible (i.e. when the menu is collapsed)
-            //so getText() will be empty.
-            //@link https://github.com/stephenharris/WordPressBehatExtension/issues/2
-            $itemName = $this->stripTagsAndContent($first_level_item->find('css', '.wp-menu-name')->getHtml());
-
-            if (strtolower($item[0]) == strtolower($itemName)) {
-                if (isset($item[1])) {
-                    $second_level_items = $first_level_item->findAll('css', 'ul li a');
-
-                    foreach ($second_level_items as $second_level_item) {
-                        $itemName = $this->stripTagsAndContent($second_level_item->getHtml());
-                        if (strtolower($item[1]) == strtolower($itemName)) {
-                            try {
-                                //Focus on the menu link so the submenu appears
-                                $first_level_item->find('css', 'a.menu-top')->focus();
-                            } catch (UnsupportedDriverActionException $e) {
-                                //This will fail for GoutteDriver but neither is it necessary
-                            }
-                            $click_node = $second_level_item;
-                            break;
-                        }
-                    }
-                } else {
-                    //We are clicking a top-level item:
-                    $click_node = $first_level_item->find('css', 'a');
-                }
-                break;
-            }
-        }
-
-        if (false === $click_node) {
-            throw new \Exception('Menu item could not be found');
-        }
-
-        $click_node->click();
+        $adminMenu = $this->adminPage->getMenu();
+        $adminMenu->clickMenuItem($item);
     }
 
     /**
@@ -123,49 +76,30 @@ class WordPressAdminContext extends RawMinkContext implements Context, SnippetAc
     public function theAdminMenuShouldAppearAs(TableNode $table)
     {
 
-        $menu_items = $this->getSession()->getPage()->findAll('css', '#adminmenu > li a .wp-menu-name');
+        $adminMenu = $this->adminPage->getMenu();
+        $topLevel = $adminMenu->getTopLevelMenuItems();
 
-        foreach ($menu_items as $index => $element) {
-            try {
-                if (! $element->isVisible()) {
-                    unset($menu_items[$index]);
-                }
-            } catch (\Exception $e) {
-                //do nothing.
-            }
+        $actualHash = array();
+        foreach ($topLevel as $actualMenuName) {
+            $actualHash[] = array( $actualMenuName );
+        }
+        $actualTableNode = new TableNode($actualHash);
+
+        if (count($topLevel) != count($table->getRows())) {
+            //var_dump( $topLevel );
+            throw new \Exception("Number of rows do not match. Found: \n" . $actualTableNode);
         }
 
-        foreach ($menu_items as $n => $element) {
-            $actual[] = array( $menu_items[$n]->getText() );
-        }
-        $actual_table = new TableNode($actual);
+        $expected = $table->getColumn(0);
 
-        try {
-            $this->assertColumnIsLike($table, $actual_table, 0);
-        } catch (\Exception $e) {
-            throw new \Exception(sprintf(
-                "Found elements:\n%s",
-                $actual_table->getTableAsString()
-            ) . "\n" . $e->getMessage());
-        }
-    }
+        foreach ($topLevel as $index => $actualMenuName) {
+            $expectedMenuName = $expected[ $index ];
 
-    protected function assertColumnIsLike($table, $actual_table, $column)
-    {
-        $expected = $table->getColumn($column);
-        $actual   = $actual_table->getColumn($column);
-
-        if (count($expected) != count($actual)) {
-            throw new \Exception('Number of rows do not match');
-        }
-
-        foreach ($actual as $row => $actual_value) {
-            $expected_value = $expected[$row];
-            if (! preg_match("/$expected_value/", $actual_value)) {
+            if (! preg_match("/$expectedMenuName/", $actualMenuName)) {
                 throw new \Exception(sprintf(
-                    'Expected "%s" but found "%s"',
-                    $expected_value,
-                    $actual_value
+                    'Expected "%s" but found "%s":' . "\n" . $actualTableNode,
+                    $expectedMenuName,
+                    $actualMenuName
                 ));
             }
         }